requirements.txt
Flask ==3.1.0
mysql-connector-python == 9.3.0
PyJWT == 2.10.1